diffusion connection recovery
Recover connection process with original state after crash.
Tell self to connect on init. this now allows the connection
to re establish the connection when restarting.

todo: consider renaming the Diffusion.Connection to Diffusion.Session as
its the gun process that is the connection, and the Diffusion one is
simply the data handler that subscribes to the connection.