api change
removed the topic handler supervisor. It's now the users
responsibility to supervise the handler. This client will now only
supervise the connection.

My reasoning for this is that the built in topic handler supervisor
might not suit all use cases. This way the client is as flexible as possible
to the users needs, they will now be able to decide whether their topic
should be permanent or temporary and the strategy.